Fix #61 - Add mapper for time format
- move all mapper files to new directory
- update test file to new directory
