+++ tests/unit/core/legacy/UserPreferencesHandlerTest.php
@@ -35,11 +35,11 @@ use Codeception\Test\Unit;
 use Exception;
 use App\Currency\LegacyHandler\CurrencyHandler;
 use App\DateTime\LegacyHandler\DateTimeHandler;
-use App\UserPreferences\LegacyHandler\UserPreferenceHandler;
-use App\UserPreferences\LegacyHandler\CurrencyPreferenceMapper;
-use App\UserPreferences\LegacyHandler\DateFormatPreferenceMapper;
-use App\UserPreferences\LegacyHandler\TimeFormatPreferenceMapper;
-use App\UserPreferences\LegacyHandler\UserPreferencesMappers;
+use App\UserPreferences\LegacyHandler\Mappers\UserPreferenceHandler;
+use App\UserPreferences\LegacyHandler\Mappers\CurrencyPreferenceMapper;
+use App\UserPreferences\LegacyHandler\Mappers\DateFormatPreferenceMapper;
+use App\UserPreferences\LegacyHandler\Mappers\TimeFormatPreferenceMapper;
+use App\UserPreferences\LegacyHandler\Mappers\UserPreferencesMappers;
 use Symfony\Component\HttpFoundation\Session\Session;
 use Symfony\Component\HttpFoundation\Session\Storage\MockArraySessionStorage;
 use User;
